Add find nearst shop method

+++ YummyFood.Domain/Entities/Address.cs
@@ -6,8 +6,8 @@ namespace YummyFood.Domain.Entities
     public class Address : BaseEntity
     {
         public string? Name { get; set; }
-        public long Letitude { get; set; }
-        public long Longitude { get; set; }
+        public float Letitude { get; set; }
+        public float Longitude { get; set; }
         public long? UserId { get; set; }
         public UserApp? User { get; set; }
     }
